requirements.txt
apispec==6.4.0
blinker==1.7.0
cachelib==0.9.0
click==8.1.7
Deprecated==1.2.14
Flask==3.0.2
Flask-Caching==2.1.0
Flask-Limiter==3.5.1
flask-smorest==0.43.0
gunicorn==21.2.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
itsdangerous==2.1.2
Jinja2==3.1.3
limits==3.8.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
marshmallow==3.20.2
mdurl==0.1.2
ordered-set==4.1.0
packaging==23.2
Pygments==2.17.2
PyJWT==2.8.0
rich==13.7.0
typing_extensions==4.9.0
webargs==8.4.0
Werkzeug==3.0.1
wrapt==1.16.0
zipp==3.17.0
